+++ onecodex/api.py
@@ -6,7 +6,6 @@ One Codex API
 """
 from __future__ import print_function
 from datetime import datetime
-import importlib
 import json
 import logging
 import os
@@ -19,16 +18,9 @@ from requests.auth import HTTPBasicAuth
 
 from onecodex.lib.auth import BearerTokenAuth
 from onecodex.models import _model_lookup
+from onecodex.utils import ModuleAlias
 
 log = logging.getLogger(__name__)
-
-
-class ModuleAlias(object):
-    def __init__(self, name):
-        self.name = name
-
-    def __repr__(self):
-        return '{} helper functions'.format(self.name)
 
 
 class Api(object):
@@ -81,14 +73,9 @@ class Api(object):
 
         # Try to import and copy key modules onto the Api object
         for module_name in ['onecodex.viz', 'onecodex.distance']:
-            try:
-                short_name = module_name.split('.')[1]
-                module = importlib.import_module(module_name)
-                setattr(self, short_name, ModuleAlias(short_name))
-                for key in module.__all__:
-                    setattr(getattr(self, short_name), key, getattr(module, key))
-            except ImportError:
-                pass
+            module = ModuleAlias(module_name)
+            if module._imported:
+                setattr(self, module._name, module)
 
     def _copy_resources(self):
         """
