+++ onecodex/viz/__init__.py
@@ -1,6 +1,6 @@
-from ._heatmap import plot_heatmap
-from ._pca import plot_pca
-from ._distance import plot_distance
-from ._metadata import plot_metadata
+from onecodex.viz._heatmap import plot_heatmap
+from onecodex.viz._pca import plot_pca
+from onecodex.viz._distance import plot_distance
+from onecodex.viz._metadata import plot_metadata
 
 __all__ = ['plot_heatmap', 'plot_pca', 'plot_distance', 'plot_metadata']
